Day 6
Quite nice. I wanted a unordered_map so I could expliot the quick hash map finding. This is what the namespace std template weird hash thing at the start was.
This is probably useful in the future for using a hash map on strange types (doesn't work automatically, and gave me a weird error about the constructor for
Guard not working at the time, so would have been knida hard to spot without some chatGPT input).

I also wanted my code to be a bit pretty so I added the direction thing, and made the guard it's own class.

Thought it would be quick in part two as we only check places that the guard went in part 1, but was still pretty slow

I also think I was lucky as I did not do anything about making sure we don't add an obstruction where the guard is.