Day 9 both parts
Fairly good. Made a bit of a mess of part 2 because I didn't understand what the effect of deleting and inserting things into
a vector would have on an existing pointer in a vector.

I claim that if you have some kind of linked list (is that the term when each item points to the location of the next one?), then if you add and erase stuff,
then a pointer looking miles away shouldn't change! It should just change the location of the previous entry's 'next' item. Oh well. we foundt the bug in the end,
and this code is pretty quick!