Day 17
Nice. Part 2 was a bit tricky. I tried to just brute force it on the train using a starting val (noticing that A had to be bigger than 8** (program length minus one possibly), but this
was still way way way too many options.

I then worked out in about 5 minutes at work, once I had calmed down a bit that we were pulling into charing cross, that we could just go backwards. The final iteration always works for any n in range(1,8),
and we know it has to be here, otherwise it wouldn't be the final iteration.

Then, we know that the previous iteration could only have been values of (8*j)+i for each j in the possible values for the final iteration, plus anything in range(8), as the floor part
of the division of A would get rid of this and gives us a valid value for the last iter, and we can check which of these values actually work by running a length of the program and seeing what is output
that time around.

We can just repeat this going up the chain, until we get the possible starting values on the first iteration! There were 2 possible values, and we return the min. Nice.