Day 20
Little bug in part two slowed me down (needed to add 1 to the top of the y range when finding
where the end of my 'cheat' could be, but it wasn't picked up by the example because the grid
was too small).

Otherwise, my approach was pretty good I think. Code runs in 5 seconds or so.